greenferries/requirements.txt
datasette
requests
wikidataintegrator
python-dotenv
openpyxl
csvs-to-sqlite
notebook
pandas
xlrd
matplotlib
awesome-slugify
db-to-sqlite[postgresql]
python-frontmatter
folium
ipywidgets
